move models to their own directory
