+++ app/assets/javascripts/models/vector.js
